Reorganize entire file structure
