+++ sdk/src/main/java/com/stytch/sdk/StytchUI.kt
@@ -1,22 +1,19 @@
 package com.stytch.sdk
 
 import android.app.Activity
-import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
-import android.os.Bundle
-import android.util.Log
+import android.content.res.Resources
+import android.graphics.Typeface
+import android.support.annotation.FontRes
 import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContract
-import com.google.android.gms.auth.api.credentials.Credential
-import com.google.android.gms.auth.api.phone.SmsRetriever
-import com.google.android.gms.common.api.CommonStatusCodes
-import com.google.android.gms.common.api.Status
-import com.wealthfront.magellan.Navigator
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
+import androidx.annotation.ColorInt
+import androidx.annotation.ColorRes
+import androidx.core.content.ContextCompat
+import androidx.core.content.res.ResourcesCompat
+import java.io.Serializable
 
 public object StytchUI {
     private var _uiCustomization: StytchUICustomization? = null
@@ -53,7 +50,7 @@ public object StytchUI {
             return context.registerForActivityResult(activityResultContract, onResult)
         }
 
-        internal val activityResultContract by lazy {
+        private val activityResultContract by lazy {
             object : ActivityResultContract<Unit, StytchUIResult>() {
                 override fun createIntent(context: Context, input: Unit): Intent {
                     return this@EmailMagicLink.createIntent(context)
@@ -91,7 +88,7 @@ public object StytchUI {
             return context.registerForActivityResult(activityResultContract, onResult)
         }
 
-        internal val activityResultContract by lazy {
+        private val activityResultContract by lazy {
             object : ActivityResultContract<Unit, StytchUIResult>() {
                 override fun createIntent(context: Context, input: Unit): Intent {
                     return this@SMSPasscode.createIntent(context)
@@ -109,104 +106,162 @@ public object StytchUI {
     }
 }
 
-internal class StytchEmailMagicLinkActivity : StytchActivity() {
+public sealed class StytchUIResult {
+    public data class Success(val user: StytchResponseTypes.AuthenticateUserResponse) : StytchUIResult()
+    public object CancelledByUser : StytchUIResult()
+}
 
-    override fun createNavigator(): Navigator {
-        if (!StytchUI.EmailMagicLink.configured) {
-            error("Stytch Error: Launch StytchUI Email Magic Link activity with first calling StytchUI.EmailMagicLink.configure(...)")
-        }
-        return Navigator.withRoot(EmailMagicLinkHomeScreen()).build()
-    }
-
-    override fun onNewIntent(intent: Intent?) {
-        super.onNewIntent(intent)
-        val token = intent?.data?.getQueryParameter("token")
-        if (token != null) {
-            // Verify deeplink
-            GlobalScope.launch {
-                val result = StytchApi.MagicLinks.authenticateMagicLink(token)
-                when (result) {
-                    is StytchResult.Success -> {
-                        Log.d("StytchLog", "Successful Magic Link Authentication")
-                        setResult(RESULT_OK, intentWithExtra(result.value))
-                        finish()
-                    }
-                    is StytchResult.Error   -> {
-                        Log.d("StytchLog", "Failed Magic Link Authentication")
-                    }
-                    StytchResult.NetworkError -> TODO()
-                }
-            }
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        when (requestCode) {
-            IntentCodes.EMAIL_PICKER_INTENT_CODE.ordinal -> {
-                if (resultCode != Activity.RESULT_OK) return
-                val emailAddress = data?.getParcelableExtra<Credential>(Credential.EXTRA_KEY)?.id
-                Log.d("StytchLog", "$emailAddress")
-                emailAddress?.let {
-                    (navigator?.currentScreen() as? EmailMagicLinkHomeScreen)?.emailAddressHintGiven(it)
-                }
-            }
+public fun Intent?.toStytchUIResult(): StytchUIResult {
+    disposeAllScreens()
+    return when (this) {
+        null -> StytchUIResult.CancelledByUser
+        else -> {
+            val user = getSerializableExtra<StytchResponseTypes.AuthenticateUserResponse>()
+                ?: error("Stytch Error: Called Intent?.toStytchUIResult on an invalid Intent. Are you checking for the correct request code in onActivityResult()?")
+            StytchUIResult.Success(user)
         }
     }
 }
 
-public class StytchSMSPasscodeActivity : StytchActivity() {
-    private val smsVerificationReceiver = object : BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-            runCatching {
-                if (intent.action == SmsRetriever.SMS_RETRIEVED_ACTION) {
-                    val extras = intent.extras
-                    if ((extras?.get(SmsRetriever.EXTRA_STATUS) as? Status)?.statusCode == CommonStatusCodes.SUCCESS) {
-                        val consentIntent = extras.getParcelable<Intent>(SmsRetriever.EXTRA_CONSENT_INTENT)
-                        startActivityForResult(consentIntent, IntentCodes.SMS_ONE_TAP_AUTOFILL_INTENT_CODE.ordinal)
-                    }
-                }
-            }
+public class StytchUICustomization(
+    public var backgroundColor: StytchColor = StytchColor.fromColorId(R.color.backgroundColor),
+    public var hideActionBar: Boolean = false,
+    public var actionBarColor: StytchColor = StytchColor.fromColorId(R.color.tertiaryBrandColor),
+    public var showTitle: Boolean = true,
+    public var titleStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.primaryBrandColor),
+        font = StytchFont.fromFontId(R.font.ibm_plex_sans, style = StytchFont.Style.BOLD),
+        size = 30.sp,
+    ),
+    public var showSubtitle: Boolean = true,
+    public var subtitleStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.primaryBrandColor),
+        size = 16.sp,
+    ),
+    public var consentTextStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.primaryBrandColor),
+        size = 14.sp,
+    ),
+    public var inputTextStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.primaryBrandColor),
+        size = 16.sp,
+    ),
+    public var inputHintStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.secondaryBrandColor),
+        size = 16.sp,
+    ),
+    public var inputBackgroundColor: StytchColor = StytchColor.fromColorId(R.color.backgroundColor),
+    public var inputBackgroundBorderColor: StytchColor = StytchColor.fromColorId(R.color.secondaryBrandColor),
+    public var inputCornerRadius: DensityIndependentPixels = 5.dp,
+    public var buttonTextStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.buttonTextColor),
+        font = StytchFont.fromFontId(R.font.ibm_plex_sans, style = StytchFont.Style.BOLD),
+        size = 16.sp,
+    ),
+    public var buttonDisabledTextColor: StytchColor = StytchColor.fromColorId(R.color.fourthBrandColor),
+    public var buttonEnabledBackgroundColor: StytchColor = StytchColor.fromColorId(R.color.primaryBrandColor),
+    public var buttonDisabledBackgroundColor: StytchColor = StytchColor.fromColorId(R.color.fifthBrandColor),
+    public var buttonCornerRadius: DensityIndependentPixels = 5.dp,
+    public var errorTextStyle: StytchTextStyle = StytchTextStyle(
+        color = StytchColor.fromColorId(R.color.errorTextColor),
+        size = 14.sp,
+    ),
+)
+
+public class StytchTextStyle(
+    public var size: ScalablePixels = 10.sp,
+    public var color: StytchColor = StytchColor.fromColorId(R.color.primaryBrandColor),
+    public var font: StytchFont = StytchFont.fromFontId(R.font.ibm_plex_sans),
+)
+
+@JvmInline
+public value class DensityIndependentPixels(private val pixels: Float) {
+    internal fun toFloat(): Float = pixels * Resources.getSystem().displayMetrics.density
+}
+
+@JvmInline
+public value class ScalablePixels(private val pixels: Float) {
+    internal fun toFloat(): Float = pixels // Resources.getSystem().displayMetrics.density // TODO adjust for font size
+}
+
+public val Number.dp: DensityIndependentPixels get() = DensityIndependentPixels(this.toFloat())
+public val Number.sp: ScalablePixels get() = ScalablePixels(this.toFloat())
+
+public class StytchColor internal constructor(
+    internal val colorId: Int? = null,
+    internal val color: Int? = null,
+) {
+
+    @ColorInt
+    public fun getColor(context: Context): Int {
+        return color ?: ContextCompat.getColor(context, colorId!!)
+    }
+
+    public companion object {
+        public fun fromColorId(@ColorRes colorId: Int): StytchColor {
+            return StytchColor(colorId = colorId)
         }
-    }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val intentFilter = IntentFilter(SmsRetriever.SMS_RETRIEVED_ACTION)
-        registerReceiver(smsVerificationReceiver, intentFilter)
-    }
-
-    override fun createNavigator(): Navigator {
-        if (!StytchUI.SMSPasscode.configured) {
-            error("Stytch Error: Launch StytchUI SMS Passcode activity with first calling StytchUI.SMSPasscode.configure(...)")
-        }
-        return Navigator.withRoot(SMSPasscodeHomeScreen()).build()
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        when (requestCode) {
-            IntentCodes.PHONE_NUMBER_PICKER_INTENT_CODE.ordinal -> {
-                if (resultCode != Activity.RESULT_OK) return
-                val phoneNumber = data?.getParcelableExtra<Credential>(Credential.EXTRA_KEY)?.id
-                phoneNumber?.let {
-                    (navigator?.currentScreen() as? SMSPasscodeHomeScreen)?.phoneNumberHintGiven(it)
-                }
-            }
-            IntentCodes.SMS_ONE_TAP_AUTOFILL_INTENT_CODE.ordinal -> {
-                if (resultCode != Activity.RESULT_OK || data == null) return
-                val smsMessage = data.getStringExtra(SmsRetriever.EXTRA_SMS_MESSAGE)
-                smsMessage?.let {
-                    (navigator?.currentScreen() as? SMSPasscodeEnterPasscodeScreen)?.smsReceived(it)
-                }
-            }
+        public fun fromColor(@ColorInt color: Int): StytchColor {
+            return StytchColor(color = color)
         }
     }
 }
 
-internal enum class IntentCodes {
-    EMAIL_PICKER_INTENT_CODE,
-    PHONE_NUMBER_PICKER_INTENT_CODE,
-    SMS_ONE_TAP_AUTOFILL_INTENT_CODE,
-    SMS_ZERO_TAP_AUTOFILL_INTENT_CODE,
+public class StytchFont private constructor(
+    @FontRes private val fontId: Int? = null,
+    private val typeface: Typeface? = null,
+    private val style: Style = Style.NORMAL,
+) {
+
+    public fun getFont(context: Context): Typeface? {
+        return if (fontId != null) {
+            val baseFont = ResourcesCompat.getFont(context, fontId)
+            when (style) {
+                Style.NORMAL -> baseFont
+                Style.BOLD -> Typeface.create(baseFont, Typeface.BOLD)
+                Style.ITALIC -> Typeface.create(baseFont, Typeface.ITALIC)
+                Style.BOLD_ITALIC -> Typeface.create(baseFont, Typeface.BOLD_ITALIC)
+            }
+        } else typeface
+    }
+
+    public enum class Style {
+        NORMAL,
+        BOLD,
+        ITALIC,
+        BOLD_ITALIC
+    }
+
+    public companion object {
+        public fun fromFontId(@FontRes fontId: Int, style: Style = Style.NORMAL): StytchFont {
+            return StytchFont(fontId = fontId, style = style)
+        }
+
+        public fun fromTypeface(typeface: Typeface?): StytchFont {
+            return StytchFont(typeface = typeface)
+        }
+    }
+}
+
+internal fun <T : StytchActivity> stytchIntent(context: Context, activity: Class<T>): Intent {
+    return Intent(context, activity).apply {
+        putExtra(StytchActivity.ACTIVITY_ID_EXTRA_NAME, StytchActivity.getNextActivityId())
+    }
+}
+
+internal inline fun <reified T : Serializable> intentWithExtra(extra: T): Intent {
+    return Intent().withSerializableExtra(extra)
+}
+
+internal inline fun <reified T : Serializable> Intent.withSerializableExtra(extra: T): Intent = apply {
+    putExtra(T::class.qualifiedName, extra)
+}
+
+internal inline fun <reified T : Serializable> Intent?.getSerializableExtra(): T? {
+    return this?.extras?.getSerializable(T::class.qualifiedName) as? T
+}
+
+internal fun disposeAllScreens() {
+    StytchActivity.navigator = null
 }
