+++ sdk/src/main/java/com/stytch/sdk/screens/EmailMagicLinkHomeScreen.kt
@@ -1,8 +1,9 @@
-package com.stytch.sdk
+package com.stytch.sdk.screens
 
 import android.content.Context
 import android.text.InputFilter
 import android.text.Spanned
+import android.util.Patterns
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
@@ -10,6 +11,16 @@ import androidx.annotation.StringRes
 import androidx.core.widget.addTextChangedListener
 import com.google.android.gms.auth.api.credentials.Credentials
 import com.google.android.gms.auth.api.credentials.HintRequest
+import com.stytch.sdk.IntentCodes
+import com.stytch.sdk.R
+import com.stytch.sdk.StytchApi
+import com.stytch.sdk.StytchEditText
+import com.stytch.sdk.StytchErrorTextView
+import com.stytch.sdk.StytchErrorTypes
+import com.stytch.sdk.StytchResult
+import com.stytch.sdk.StytchScreen
+import com.stytch.sdk.StytchScreenView
+import com.stytch.sdk.StytchUI
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -157,34 +168,6 @@ internal class EmailMagicLinkHomeView(context: Context) : StytchScreenView<Email
     }
 }
 
-internal class EmailMagicLinkConfirmationScreen(
-    private val emailAddress: String,
-) : StytchScreen<EmailMagicLinkConfirmationView>() {
-    override fun createView(context: Context): EmailMagicLinkConfirmationView {
-        return EmailMagicLinkConfirmationView(context).apply {
-            setBackgroundColor(StytchUI.uiCustomization.backgroundColor.getColor(context))
-
-            description.text = resources.getString(R.string.magic_link_sent_description, emailAddress)
-
-            tryAgainButton.setOnClickListener {
-                navigator.goBack()
-            }
-        }
-    }
-}
-
-internal class EmailMagicLinkConfirmationView(context: Context) : StytchScreenView<EmailMagicLinkConfirmationScreen>(context) {
-    val title: TextView
-    val description: TextView
-    val tryAgainButton: Button
-
-    init {
-        inflate(context, R.layout.magic_link_check_your_email_layout, this)
-        title = findViewById(R.id.title)
-        description = findViewById(R.id.description)
-        tryAgainButton = findViewById(R.id.try_again_button)
-    }
-
-    override fun subscribeToState() {
-    }
+internal fun String.isValidEmailAddress(): Boolean {
+    return Patterns.EMAIL_ADDRESS.matcher(this).matches()
 }
